CrimSonhelper
crimsonhelper
crimsonhelper helps with kuudras quest and it shows all the locations of the npcs in crimson isle

PLEASE NOTE: This is a forge mod. Not a spigot plugin. Use the forge mod loader to use this mod for yourself.

Usage
Download the jar file from the top of the page and place it into your minecraft installation's mod folder (usually %appdata%/.minecraft/mods)
Start minecraft with forge and enjoy!

How to use
You can configure the mod in the mods menu. Check your forge installaton on where to find the settings panel

Commands
/kf - Opens the mod menu with a guide on how to use the mod effectively

Source code
If you would like to help us out and contribute to the mod, please kindly check our Github repository at Github.